separate spoonacular_api_impl into its own file
